+++ apps/cli/cli_autocli.h
@@ -52,6 +52,7 @@ enum autocli_op{
  */
 int autocli_module(clicon_handle h, char *modname, int *enable);
 int autocli_completion(clicon_handle h, int *completion);
+int autocli_grouping_treeref(clicon_handle h, int *grouping_treeref);
 int autocli_list_keyword(clicon_handle h, autocli_listkw_t *listkw);
 int autocli_compress(clicon_handle h, yang_stmt *ys, int *compress);
 int autocli_treeref_state(clicon_handle h, int *treeref_state);