+++ apps/cli/cli_generate.h
@@ -48,6 +48,11 @@
  * where the virtual callback (overwrite_me) is overwritten by cli_set.
  */
 #define GENERATE_CALLBACK "overwrite_me"
+#define GROUPING_CALLBACK "prepend_me"
+#define MTPOINT_PREFIX    "mtpoint:"
+
+/* variable expand function */
+#define GENERATE_EXPAND_XMLDB "expand_dbvar"
 
 /* Name of autocli CLIgen treename
  */